# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mahad Nafsy API",
    "version": "2.0.0",
    "description": "Auto-generated from api/routes/*.js by tools/gen-openapi.mjs. Operation bodies are stubs; enrich summaries/schemas as needed."
  },
  "servers": [
    {
      "url": "https://mahadnafsy.com"
    },
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/admin/migrate-branches": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "POST /api/admin/migrate-branches",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/admin/payments": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "GET /api/admin/payments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/admin/payments/review": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "GET /api/admin/payments/review",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}